https://github.com/Glitchfiend/BiomesOPlenty/commit/4d52f9a7bd03aa3f6a5497c4d394b85a30a24e46 & convert nextDouble() -> nextFloat() & optimize gen WorldGenSplotches & etc

+++ src/main/java/biomesoplenty/api/biome/BOPBiomeDecorator.java
@@ -95,7 +95,7 @@ public class BOPBiomeDecorator<T extends BiomeFeatures> extends BiomeDecorator
             completeWeight += Double.parseDouble(String.valueOf(weight));
         }
 
-        double random = (new XSTR()).nextDouble() * completeWeight;
+        double random = (new XSTR()).nextFloat() * completeWeight;
         double countWeight = 0D;
 
         for (Map.Entry<T, ? extends Number> entry : worldGeneratorMap.entrySet())
