+++ src/main/java/biomesoplenty/client/particles/EntityDandelionFX.java
@@ -22,10 +22,10 @@ public class EntityDandelionFX extends EntityFX
 		motionY *= 0.10000000149011612D;
 		motionZ *= 0.20000000149011612D;
 
-		float f4 = (float)(new XSTR()).nextDouble() * 0.4F + 0.6F;
+		float f4 = (new XSTR()).nextFloat() * 0.4F + 0.6F;
 		particleScale *= 0.25F;
 		particleScale *= par8;
-		particleMaxAge = (int)(8.0D / ((new XSTR()).nextDouble() * 0.8D + 0.2D));
+		particleMaxAge = (int)(8.0F / ((new XSTR()).nextFloat() * 0.8F + 0.2F));
 		particleMaxAge = (int)(particleMaxAge * par8);
 		noClip = false;
