+++ src/main/java/biomesoplenty/common/entities/EntityPixie.java
@@ -100,7 +100,7 @@ public class EntityPixie extends EntityFlyingCreature
     	    {
     	        if (this.rand.nextInt(2) == 0)
     	        {
-    	            BiomesOPlenty.proxy.spawnParticle("pixietrail", this.posX + (this.rand.nextDouble()) * (double)this.width, this.posY + this.rand.nextDouble() * (double)this.height - (double)this.yOffset, this.posZ + (this.rand.nextDouble()) * (double)this.width);
+    	            BiomesOPlenty.proxy.spawnParticle("pixietrail", this.posX + (this.rand.nextFloat()) * (double)this.width, this.posY + this.rand.nextFloat() * (double)this.height - (double)this.yOffset, this.posZ + (this.rand.nextFloat()) * (double)this.width);
     	        }
     	    }
     	}
